+++ sources/ocf_topics.php
@@ -73,12 +73,12 @@ function render_topic_box($row,$zone='_SEARCH',$give_context=true,$include_bread
  */
 function ocf_get_topic_where($topic_id)
 {
-	$where='p_topic_id='.strval((integer)$topic_id);
+	$where='p_topic_id='.strval($topic_id);
 	if (is_guest())
 		$where.=' AND p_intended_solely_for IS NULL';
 	elseif (!has_privilege(get_member(),'view_other_pt'))
-		$where.=' AND (p_intended_solely_for='.strval((integer)get_member()).' OR p_poster='.strval((integer)get_member()).' OR p_intended_solely_for IS NULL)';
-	if ((!has_privilege(get_member(),'see_unvalidated')) && (addon_installed('unvalidated'))) $where.=' AND (p_validated=1 OR ((p_poster<>'.strval($GLOBALS['FORUM_DRIVER']->get_guest_id()).' OR '.db_string_equal_to('p_ip_address',get_ip_address()).') AND p_poster='.strval((integer)get_member()).'))';
+		$where.=' AND (p_intended_solely_for='.strval(get_member()).' OR p_poster='.strval(get_member()).' OR p_intended_solely_for IS NULL)';
+	if ((!has_privilege(get_member(),'see_unvalidated')) && (addon_installed('unvalidated'))) $where.=' AND (p_validated=1 OR ((p_poster<>'.strval($GLOBALS['FORUM_DRIVER']->get_guest_id()).' OR '.db_string_equal_to('p_ip_address',get_ip_address()).') AND p_poster='.strval(get_member()).'))';
 	return $where;
 }
